chore: split common.ts to folder
