+++ src/components/SkuOptions/SkuOptions.ts
@@ -2,7 +2,7 @@ import { assertRequired } from "@/utils/assertRequired"
 import { intersectionOf } from "@/utils/intersectionOf"
 import { injectStore, Store } from "../Product/store"
 import { customElement } from "../decorators"
-import { syncSkuData } from "../common"
+import { syncSkuData } from "../common/syncSkuData"
 import { NostoElement } from "../Element"
 
 /**
