+++ src/components/common/getTemplate.ts
@@ -1,14 +1,3 @@
-import { Store } from "./Product/store"
-
-export function syncSkuData(element: HTMLElement, setter: Store["setSkuFields"]) {
-  setter({
-    image: element.getAttribute("n-img") ?? undefined,
-    altImage: element.getAttribute("n-alt-img") ?? undefined,
-    price: element.getAttribute("n-price") ?? undefined,
-    listPrice: element.getAttribute("n-list-price") ?? undefined
-  })
-}
-
 interface CampaignElement extends HTMLElement {
   template?: string
   templateElement?: HTMLTemplateElement
